# PERF-#5575: Don't trigger axes computation in `reset_index` function