--- a/modin/core/dataframe/pandas/partitioning/partition.py
+++ b/modin/core/dataframe/pandas/partitioning/partition.py
@@ -315,7 +315,7 @@ class PandasDataframePartition(ABC):  # pragma: no cover
         ----------
         materialize : bool, default: True
             Whether to forcibly materialize the result into an integer. If ``False``
-            was specified may return a future of the result if it hasn't been
+            was specified, may return a future of the result if it hasn't been
             materialized yet.
 
         Returns
@@ -335,7 +335,7 @@ class PandasDataframePartition(ABC):  # pragma: no cover
         ----------
         materialize : bool, default: True
             Whether to forcibly materialize the result into an integer. If ``False``
-            was specified may return a future of the result if it hasn't been
+            was specified, may return a future of the result if it hasn't been
             materialized yet.
 
         Returns