@anmyachev 2023-01-26
> overall looks good!
> 
> The only thing I'm worried a bit is that you moved the pandas-specific validity check to the query compiler from the high-level API level. It's an upsetting trend that we're constantly moving pandas-depending stuff to the query compiler thus losing the layers' separatability.
> 
> An initial sense of `modin.pandas` layer was sanitizing inputs/preprocessing parameters/catching errors thus being able to pass to the QC much more standardized and clean from error inputs. Now it seems that the role of matching pandas behavior goes more and more to the backend's responsibility which is quite sad and certainly doesn't scale well with the philosophy of having various amounts of execution backends in modin. Hope we would find a way to gently delegate these error checkings to the backend without making them know a lot about pandas and not losing in performance at the same time. But of course, it's a topic for a separate discussion.

The main reason is that the frontend code is not designed for asynchronous execution.

In Modin we already have the following code:

```python
# avoid materializing columns in lazy mode. the query compiler
# will handle errors where dtype dict includes keys that are not
# in columns.
if (
    not self._query_compiler.lazy_execution
    and not set(dtype.keys()).issubset(set(self._query_compiler.columns))
    and errors == "raise"
):
    raise KeyError(
        "Only a column name can be used for the key in "
        + "a dtype mappings argument."
    )
```

Perhaps I can do the same for the check. However, in the future, this will lead to the fact that there will be at least 2 code branches for each action: asynchronous and synchronous.