@dchigarev 2023-01-26
> Perhaps I can do the same for the check.

I'm okay with leaving things as they are in this PR for now. Just wanted to express my soul pain regarding this trend :D