@anmyachev 2023-01-27
A full axis implementation of `reset_index` can bring slowdowns even if the execution is asynchronous. We should use the map operator. To do this, I will also need to transfer index lengths asynchronously in order to correctly create `RangeIndex` in each partition. Found by @dchigarev, thanks!