@anmyachev 2023-01-30
@dchigarev your implementation of `reset_index` works well for my workload.