@anmyachev 2023-02-03
> Is there any comparison of the changes against master?

Version | Sync mode | Async mode
-|-|-
Master 1rst |  4.09 |  1.91
Master 2nd |  0.98 |  0.23
Master 3rd |  0.81 | 0.13
PR 1rst | 4.01 | 0.05
PR 2nd | 1.94 | 0.06
PR 3rd | 1.83 | 0.06

<details>
<summary> Benchmark: Ray, 32 cores </summary>

```bash
import modin.pandas as pd
import numpy as np
from asv_bench.benchmarks.utils.common import execute
from time import time

data = np.random.randint(10000, size=(1000000, 1000))
df = pd.DataFrame(data)

for _ in range(3):
    df._query_compiler._modin_frame._index_cache = None
    start = time()
    res = df.reset_index()
    execute(res)
    print(f"reset_index time: {time()-start}")
```

</details>