@dchigarev 2023-02-03
it's strange that the PR is slower in sync mode...