@anmyachev 2023-02-03
> it's strange that the PR is slower in sync mode...
> 
> Am I understanding correctly that 'sync' mode is `execute(res)` and 'async' is without calling `execute()`?

Yes.
Why strange? We are doing a full axis operation.