@dchigarev 2023-02-03
> Why strange? We are doing a full axis operation.

Right, the more column partitions we have the slower the new implementation will be. Here are some more measurements in sync mode with the same script but with different shapes:
NPartitions = 112, the table shows time of the 3rd iteration:
| | Master | PR |
|-|-|-|
|(1_000_000, 10) | 0.13s | 0.06s |
|(1_000_000, 32) | 0.16s | 0.07s|
|(1_000_000, 256) | 0.51s | 0.45s |
|(1_000_000, 512) | 1.01s | 0.99s |
|(1_000_000, 1024) | 1.41s | 1.84s |
|(10_000_000, 32) | 1.38s | 0.55s |
|(10_000_000, 64) | 1.37s | 1.11s |