@dchigarev 2023-02-03
Here are also measurements in sync mode, but here we're triggering an '.index' materialization at the end of '.reset_index':
| | Master | PR |
|-|-|-|
|(1_000_000, 10) | 0.11s | 0.1s |
|(1_000_000, 32) | 0.13s | 0.09s|
|(1_000_000, 256) | 0.14s | 0.45s |
|(1_000_000, 512) | 0.16s | 0.98s |
|(1_000_000, 1024) | 0.18s | 1.93s |
|(10_000_000, 32) | 1.36s | 0.49s |
|(10_000_000, 64) | 1.53s | 0.96s |

<details><summary>script</summary>

```python
import modin.pandas as pd
import numpy as np
from asv_bench.benchmarks.utils.common import execute
from time import time

sizes = [
    (1_000_000, 10),
    (1_000_000, 32),
    (1_000_000, 256),
    (1_000_000, 512),
    (1_000_000, 1024),
    (10_000_000, 32),
    (10_000_000, 64),
]

for size in sizes:
    data = np.random.randint(10000, size=size)
    df = pd.DataFrame(data)

    for _ in range(3):
        df._query_compiler._modin_frame._index_cache = None
        start = time()
        res = df.reset_index()
        len(res.index) # trigger index materialization
        t_Res = time()-start
    print(size, t_Res)
```

</details>