@YarShev 2023-02-03
Should we have some dispatching mechanism depending on the number of columns?