@YarShev 2023-02-03
shape | Master | PR
-- | -- | --
(1_000_000, 256) | 0.14s | 0.45s
(1_000_000, 512) | 0.16s | 0.98s
(1_000_000, 1024) | 0.18s | 1.93s

Based on the last comment, it doesn't look very good.