@anmyachev 2023-02-03
> Should we have some dispatching mechanism depending on the number of columns?

For the workload, which I accelerated, an asynchronous implementation of this operation is needed for a dataframe with ~250 columns.