@YarShev 2023-02-06
Even though wide frames are not a common use case.