@anmyachev 2023-02-07
> I think we should detect the limits of `row_lengths` and `column_widths` using which we get slower. Depending on the limits we would use either previous or the new implemtation. Can you create an issue related to that?

https://github.com/modin-project/modin/issues/5633